requirements.txt
azure-storage-file-datalake
azure-identity
azure-accounts
validators
deltalake
pandas
pyodbc
pyarrow